Fix odr_cal_reg_set/get API, change type uint8 to int8

+++ asm330lhbg1_reg.c
@@ -1714,7 +1714,7 @@ int32_t asm330lhbg1_fifo_out_raw_get(const stmdev_ctx_t *ctx, uint8_t *val)
   * @retval        Interface status (MANDATORY: return 0 -> no Error).
   *
   */
-int32_t asm330lhbg1_odr_cal_reg_set(const stmdev_ctx_t *ctx, uint8_t val)
+int32_t asm330lhbg1_odr_cal_reg_set(const stmdev_ctx_t *ctx, int8_t val)
 {
   asm330lhbg1_internal_freq_fine_t internal_freq_fine;
   int32_t ret;
@@ -1723,7 +1723,7 @@ int32_t asm330lhbg1_odr_cal_reg_set(const stmdev_ctx_t *ctx, uint8_t val)
                              (uint8_t *)&internal_freq_fine, 1);
   if (ret == 0)
   {
-    internal_freq_fine.freq_fine = (uint8_t)val;
+    internal_freq_fine.freq_fine = val;
     ret = asm330lhbg1_write_reg(ctx, ASM330LHBG1_INTERNAL_FREQ_FINE,
                                 (uint8_t *)&internal_freq_fine, 1);
   }
@@ -1740,7 +1740,7 @@ int32_t asm330lhbg1_odr_cal_reg_set(const stmdev_ctx_t *ctx, uint8_t val)
   * @retval        Interface status (MANDATORY: return 0 -> no Error).
   *
   */
-int32_t asm330lhbg1_odr_cal_reg_get(const stmdev_ctx_t *ctx, uint8_t *val)
+int32_t asm330lhbg1_odr_cal_reg_get(const stmdev_ctx_t *ctx, int8_t *val)
 {
   asm330lhbg1_internal_freq_fine_t internal_freq_fine;
   int32_t ret;
@@ -1752,7 +1752,7 @@ int32_t asm330lhbg1_odr_cal_reg_get(const stmdev_ctx_t *ctx, uint8_t *val)
     return ret;
   }
 
-  *val = internal_freq_fine.freq_fine;
+  *val = (int8_t)internal_freq_fine.freq_fine;
 
   return ret;
 }
